% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ccSpill_folders.R
\name{create_ccSpill_folders}
\alias{create_ccSpill_folders}
\title{Create a new project from the ccSpill template}
\usage{
create_ccSpill_folders(path)
}
\arguments{
\item{path}{The name of the new project directory to be created.}
}
\description{
This function will copy the 'my_project_template' folder structure
into a new directory on your computer, so you can start a
new analysis.
}
\examples{
\dontrun{
# Creates a new folder named "MyAnalysis" in your current directory
create_ccSpill_project("MyAnalysis")
}
}
