% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_as_isotopes.R
\name{rename_as_isotopes}
\alias{rename_as_isotopes}
\title{Rename Columns Based on Numeric Suffix}
\usage{
rename_as_isotopes(col_names_vector)
}
\arguments{
\item{col_names_vector}{A character vector of column names or strings.
(e.g., \code{c("X.123.999", "Y.456.998")})}
}
\value{
A character vector of the same length, containing the
incremented numbers as strings. (e.g., \code{c("124", "457")})
}
\description{
Replaces the exact isotope number with the rounded isotope number by
extracting the first numeric block following a period (\code{.}) in a character
vector, incrementing this number by 1, and returning the result as a new
character vector.
}
\details{
This function requires the 'stringr' package.

The regex \verb{(?<=\\\\.)\\\\d+} specifically looks for a sequence of digits (\verb{\\\\d+})
that is preceded by a literal dot (\verb{(?<=\\\\.)}).
}
\examples{
\dontrun{
  old_names <- c("some.name.123", "another.456", "no_number", "X.789")
  new_names <- rename_as_isotopes(old_names)
  # new_names would be: c("124", "457", NA_character_, "790")
  print(new_names)
}

}
